Fix compilation bug in optiboot

+++ Firmware/optiboot_w25x20cl.cpp
@@ -126,7 +126,7 @@ uint8_t optiboot_w25x20cl_enter()
       // i.e. rx_buffer.head == SerialHead would not be checked at all!
       // With the volatile keyword the compiler generates exactly the same code as without it with only one difference:
       // the last brne instruction jumps onto the (*rx_head == SerialHead) check and NOT onto the wdr instruction bypassing the check.
-      volatile int *rx_head = &rx_buffer.head;
+      volatile uint8_t *rx_head = &rx_buffer.head;
       while (*rx_head == SerialHead) {
         wdt_reset();
         if ( --boot_timer == 0) {
